refactor: 整理所有算子
