+++ src/03onnx/src/infer/mat_mul.cpp
@@ -1,0 +1,58 @@
+﻿#include "infer.h"
+
+namespace refactor::onnx {
+    using namespace refactor::common;
+
+    InferResult inferMatMul(Operator const &op, Tensors inputs) {
+        EXPECT_SIZE(2) {
+            auto const &a = inputs[0];
+            auto const &b = inputs[1];
+            auto dataType = a->dataType;
+            if (!isNumbericDataType(dataType) || b->dataType != dataType) {
+                return Err(InferError(ERROR_MSG("Input data type not support")));
+            }
+            auto sa = a->shape, sb = b->shape;
+            switch (sa.size()) {
+                case 1:
+                    sa.insert(sa.begin(), DimExpr(1));
+                    break;
+                case 0:
+                    return Err(InferError(ERROR_MSG("Input shape not support")));
+                default:
+                    break;
+            }
+            switch (sb.size()) {
+                case 1:
+                    sb.emplace_back(1);
+                    break;
+                case 0:
+                    return Err(InferError(ERROR_MSG("Input shape not support")));
+                default:
+                    break;
+            }
+            auto k = *sa.rbegin();
+            sa.pop_back();
+            auto m = *sa.rbegin();
+            sa.pop_back();
+            auto n = *sb.rbegin();
+            sb.pop_back();
+            if (k != *sb.rbegin()) {
+                return Err(InferError(ERROR_MSG("Input shape not support")));
+            }
+            sb.pop_back();
+            if (!sa.empty() || !sb.empty()) {
+                auto res = multidirBroadcast({sa, sb});
+                if (res.isErr()) {
+                    return Err(InferError(ERROR_MSG(res.unwrapErr())));
+                } else {
+                    auto shape = res.unwrap();
+                    shape.emplace_back(m);
+                    shape.emplace_back(n);
+                    return Ok(Tensors{std::make_shared<Tensor>(dataType, shape)});
+                }
+            } else {
+                return Ok(Tensors{std::make_shared<Tensor>(dataType, Shape{m, n})});
+            }
+        }
+    }
+}// namespace refactor::onnx
