+++ src/03onnx/src/infer/where.cpp
@@ -28,34 +28,16 @@ namespace refactor::onnx {
             auto eleSize = dataTypeSize(dataType);
             auto blob = std::make_shared<Blob>(new uint8_t[size * eleSize]);
             auto dst = reinterpret_cast<uint8_t *>(blob->ptr);
-            auto srcC = reinterpret_cast<bool *>(condition->data->ptr);
-            auto srcX = reinterpret_cast<uint8_t *>(x->data->ptr);
-            auto srcY = reinterpret_cast<uint8_t *>(y->data->ptr);
             fmt::print("( {} dst<{}> = ", op.opType.name(), size);
             for (size_t i = 0; i < size; ++i) {
                 auto indices = buildIndices(shape, i);
-                auto ptr = [&indices](Tensor const &input) -> uint8_t * {
-                    auto it0 = indices.rbegin(),
-                         end0 = indices.rend();
-                    auto it1 = input.shape.rbegin(),
-                         end1 = input.shape.rend();
-                    size_t ii = 0, mul = 1;
-                    while (it0 != end0 && it1 != end1) {
-                        ii += *it0++ * mul;
-                        mul *= it1++->value();
-                    }
-                    auto src = reinterpret_cast<uint8_t *>(input.data->ptr);
-                    return reinterpret_cast<uint8_t *>(src + ii * dataTypeSize(input.dataType));
-                };
-
-                if (*reinterpret_cast<bool *>(ptr(*condition))) {
-                    std::copy_n(ptr(*x), eleSize, dst);
+                if (*reinterpret_cast<bool *>(locate(*condition, indices))) {
+                    std::memcpy(dst + i * eleSize, locate(*x, indices), eleSize);
                     fmt::print("x ");
                 } else {
-                    std::copy_n(ptr(*y), eleSize, dst);
+                    std::memcpy(dst + i * eleSize, locate(*y, indices), eleSize);
                     fmt::print("y ");
                 }
-                dst += eleSize;
             }
             fmt::print(")");
             return Ok(Tensors{std::make_shared<Tensor>(dataType, std::move(output), std::move(blob))});
